correct data-type for days, sort users by name

+++ src/main/java/com/topdesk/yanap/web/SprintAndUsers.java
@@ -18,7 +18,7 @@ public class SprintAndUsers {
 	class UserAndDays {
 		long id;
 		String name;
-		List<Integer> days;
+		List<Float> days;
 	}
 	long id;
 	String name;
